Installs ansible from packages by default.
(required by RedHat support.)

+++ scripts/ansibleconfigserver_bootstrap.sh
@@ -55,12 +55,17 @@ qs_retry_command 10 yum -y install atomic-openshift-excluder atomic-openshift-do
 qs_retry_command 10 yum install -y https://s3-us-west-1.amazonaws.com/amazon-ssm-us-west-1/latest/linux_amd64/amazon-ssm-agent.rpm
 systemctl start amazon-ssm-agent
 systemctl enable amazon-ssm-agent
-CURRENT_PLAYBOOK_VERSION=https://github.com/openshift/openshift-ansible/archive/openshift-ansible-${OCP_ANSIBLE_RELEASE}.tar.gz
-curl  --retry 5  -Ls ${CURRENT_PLAYBOOK_VERSION} -o openshift-ansible.tar.gz
-tar -zxf openshift-ansible.tar.gz
-rm -rf /usr/share/ansible
-mkdir -p /usr/share/ansible
-mv openshift-ansible-* /usr/share/ansible/openshift-ansible
+
+if [ "${GET_ANSIBLE_FROM_GIT}" == "True" ]
+  CURRENT_PLAYBOOK_VERSION=https://github.com/openshift/openshift-ansible/archive/openshift-ansible-${OCP_ANSIBLE_RELEASE}.tar.gz
+  curl  --retry 5  -Ls ${CURRENT_PLAYBOOK_VERSION} -o openshift-ansible.tar.gz
+  tar -zxf openshift-ansible.tar.gz
+  rm -rf /usr/share/ansible
+  mkdir -p /usr/share/ansible
+  mv openshift-ansible-* /usr/share/ansible/openshift-ansible
+else
+  qs_retry_command 10 yum -y install openshift-ansible
+fi
 
 qs_retry_command 10 yum -y install atomic-openshift-excluder atomic-openshift-docker-excluder
 atomic-openshift-excluder unexclude
